Initial commit of CreateTopoTransfer.py
This file creates a basic setup for Houdini's Topo Transfer node with required dependencies and prompts the user for the file path to the source mesh to be retopologized.  In addition, it handles the setup of unified scaling for the source and target meshes for ease of use.  The script utilizes the template head geometry as the target mesh available within Houdini.